Add GstBufferMapGuard class for safer buffer management
